sync: refactor syncer code to prepare for quic syncer (#256)
Reorganizes syncer code into more modules to prepare for adding the quic
syncer:
- `task.rs` - responsible for periodically syncing with peers
- `tcp.rs` - the TCP client/server syncer implementation
- `actions.rs` - the graph actions/effects API

Future work:
Implement QUIC syncer in new `quic.rs` module.

---------
