+++ crates/aranya-daemon/src/api.rs
@@ -31,11 +31,11 @@ use tokio::{
 use tracing::{debug, error, info, instrument, trace, warn};
 
 use crate::{
+    actions::Actions,
     aqc::Aqc,
-    aranya::Actions,
     daemon::KS,
     policy::{ChanOp, Effect, KeyBundle, Role},
-    sync::SyncPeers,
+    sync::task::SyncPeers,
     Client, EF,
 };
 
@@ -57,8 +57,8 @@ pub(crate) struct DaemonApiServer {
     sk: ApiKey<CS>,
     /// The UDS path we serve the API on.
     uds_path: PathBuf,
-
-    client: Arc<Client>,
+    /// The Aranya client.
+    client: Client,
     /// The local network address for the `Client`'s sync server.
     local_addr: SocketAddr,
     /// Public keys of current device.
@@ -76,7 +76,7 @@ impl DaemonApiServer {
     #[instrument(skip_all)]
     #[allow(clippy::too_many_arguments)]
     pub(crate) fn new(
-        client: Arc<Client>,
+        client: Client,
         local_addr: SocketAddr,
         uds_path: PathBuf,
         sk: ApiKey<CS>,
@@ -240,7 +240,7 @@ impl EffectHandler {
 /// (inside [`Api`]).
 #[derive(Debug)]
 struct ApiInner {
-    client: Arc<Client>,
+    client: Client,
     /// Local socket address of the API.
     local_addr: SocketAddr,
     /// Public keys of current device.
