+++ crates/aranya-daemon/src/lib.rs
@@ -30,17 +30,18 @@
     unused_qualifications
 )]
 
+pub mod actions;
 pub mod aranya;
 pub mod config;
 #[rustfmt::skip]
 pub mod policy;
+pub mod sync;
 pub mod vm_policy;
 
 mod api;
 mod aqc;
 mod daemon;
 mod keystore;
-mod sync;
 
 pub use daemon::*;
 pub use keystore::AranyaStore;
